+++ packages/storybook/src/nimble/anchor-step/anchor-step-matrix.stories.ts
@@ -7,13 +7,18 @@ import { iconCogTag } from '@ni/nimble-components/dist/esm/icons/cog';
 import {
     createMatrix,
     sharedMatrixParameters,
-    cartesianProduct,
-    createMatrixInteractionsFromStates
+    createMatrixInteractions
 } from '../../utilities/matrix';
 import { createFixedThemeStory, createStory } from '../../utilities/storybook';
 import { hiddenWrapper } from '../../utilities/hidden';
-import { selectedStates, severityStates, stepContentStateShort, stepContentStateStepIndicator, stepContentStates, stepLayoutStates, type SelectedState, type SeverityStates, type StepContentStates, type StepLayoutStates } from '../stepper/types';
-import { backgroundStates, disabledStates, type DisabledState } from '../../utilities/states';
+import {
+    selectedStates, type SelectedState,
+    severityStates, type SeverityStates,
+    stepContentStates, stepContentStateShort, stepContentStateStepIndicator, type StepContentStates,
+    stepLayoutStates, type StepLayoutStates,
+    stepManipulationState, type StepManipulationState,
+} from '../stepper/types';
+import { backgroundStates } from '../../utilities/states';
 
 const metadata: Meta = {
     title: 'Tests/Anchor Step',
@@ -36,18 +41,18 @@ if (remaining.length > 0) {
 export default metadata;
 
 const component = (
-    [disabledName, disabled]: DisabledState,
+    [manipulationName, readonly, disabled]: StepManipulationState,
     [layoutName, isLast, orientation]: StepLayoutStates,
     [selectedName, selected]: SelectedState,
     [severityName, severity]: SeverityStates,
     [contentName, titleContent, subtitleContent, severityTextContent, stepIndicatorVisible]: StepContentStates,
 ): ViewTemplate => html`
     <div style="display: inline-flex; flex-direction: column; width: min-content;">
-        <div style="padding-right: 16px;">${disabledName} ${selectedName} Severity(${severityName}) Layout(${layoutName}) Content(${contentName})</div>
+        <div style="padding-right: 16px;">${manipulationName} ${selectedName} Severity(${severityName}) Layout(${layoutName}) Content(${contentName})</div>
         <${stepperTag} style="padding-bottom: 16px; width: 200px; height: 80px;" orientation="${() => orientation}">
             ${repeat(() => [isLast, !isLast], html`
                 <${anchorStepTag} href="#" target="_self"
-                    ?disabled=${() => disabled} ?selected=${() => selected} severity-text="${() => severityTextContent}" severity="${() => severity}"
+                    ?disabled=${() => disabled} ?readonly=${() => readonly} ?selected=${() => selected} severity-text="${() => severityTextContent}" severity="${() => severity}"
                     style="${currentIsLast => (currentIsLast ? 'display:none;' : '')}">
                     ${when(() => titleContent !== undefined, html`<span slot="title">${() => titleContent}</span>`)}
                     ${when(() => subtitleContent !== undefined, html`<span slot="subtitle">${() => subtitleContent}</span>`)}
@@ -58,7 +63,7 @@ const component = (
     </div>
 `;
 
-const matrixTemplate = html`
+const matrixTemplate = (template: ViewTemplate): ViewTemplate => html`
     <div style="
         display: grid;
         grid-template-columns: ${'1fr '.repeat(stepContentStates.length)};
@@ -66,46 +71,273 @@ const matrixTemplate = html`
         color: var(${bodyFontColor.cssCustomProperty});
         width: 1200px;
     ">
-    ${createMatrix(component, [
-        disabledStates,
+    ${template}
+    </div>
+`;
+
+export const matrix$LightTheme$ReadOnlyAbsent$DisabledAbsent: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.none],
         stepLayoutStates,
         selectedStates,
         severityStates,
         stepContentStates,
-    ])}
-    </div>
-`;
+    ])),
+    lightThemeWhiteBackground
+);
 
-export const matrix$LightTheme: StoryFn = createFixedThemeStory(matrixTemplate, lightThemeWhiteBackground);
-export const matrix$DarkTheme: StoryFn = createFixedThemeStory(matrixTemplate, darkThemeBlackBackground);
-export const matrix$ColorTheme: StoryFn = createFixedThemeStory(matrixTemplate, colorThemeDarkGreenBackground);
+export const matrix$LightTheme$ReadOnlyAbsent$Disabled: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.disabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    lightThemeWhiteBackground
+);
 
-const interactionStatesHover = cartesianProduct([
-    disabledStates,
-    stepLayoutStates,
-    selectedStates,
-    severityStates,
-    [stepContentStateShort, stepContentStateStepIndicator],
-] as const);
+export const matrix$LightTheme$ReadOnly$DisabledAbsent: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.readOnly],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    lightThemeWhiteBackground
+);
 
-const interactionStates = cartesianProduct([
-    disabledStates,
-    stepLayoutStates,
-    selectedStates,
-    severityStates,
-    [stepContentStateShort, stepContentStateStepIndicator],
-] as const);
+export const matrix$LightTheme$ReadOnly$Disabled: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.readOnlyDisabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    lightThemeWhiteBackground
+);
 
-const interactionsTemplate = createMatrixInteractionsFromStates(component, {
-    hover: interactionStatesHover,
-    hoverActive: interactionStates,
-    active: interactionStates,
-    focus: interactionStates
-});
+export const matrix$DarkTheme$ReadOnlyAbsent$DisabledAbsent: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.none],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    darkThemeBlackBackground
+);
 
-export const interactions$LightTheme: StoryFn = createFixedThemeStory(interactionsTemplate, lightThemeWhiteBackground);
-export const interactions$DarkTheme: StoryFn = createFixedThemeStory(interactionsTemplate, darkThemeBlackBackground);
-export const interactions$ColorTheme: StoryFn = createFixedThemeStory(interactionsTemplate, colorThemeDarkGreenBackground);
+export const matrix$DarkTheme$ReadOnlyAbsent$Disabled: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.disabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    darkThemeBlackBackground
+);
+
+export const matrix$DarkTheme$ReadOnly$DisabledAbsent: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.readOnly],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    darkThemeBlackBackground
+);
+
+export const matrix$DarkTheme$ReadOnly$Disabled: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.readOnlyDisabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    darkThemeBlackBackground
+);
+
+export const matrix$ColorTheme$ReadOnlyAbsent$DisabledAbsent: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.none],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    colorThemeDarkGreenBackground
+);
+
+export const matrix$ColorTheme$ReadOnlyAbsent$Disabled: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.disabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    colorThemeDarkGreenBackground
+);
+
+export const matrix$ColorTheme$ReadOnly$DisabledAbsent: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.readOnly],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    colorThemeDarkGreenBackground
+);
+
+export const matrix$ColorTheme$ReadOnly$Disabled: StoryFn = createFixedThemeStory(
+    matrixTemplate(createMatrix(component, [
+        [stepManipulationState.readOnlyDisabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        stepContentStates,
+    ])),
+    colorThemeDarkGreenBackground
+);
+
+export const interactions$LightTheme$ReadOnlyAbsent$DisabledAbsent: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.none],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    lightThemeWhiteBackground
+);
+
+export const interactions$LightTheme$ReadOnlyAbsent$Disabled: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.disabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    lightThemeWhiteBackground
+);
+
+export const interactions$LightTheme$ReadOnly$DisabledAbsent: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.readOnly],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    lightThemeWhiteBackground
+);
+
+export const interactions$LightTheme$ReadOnly$Disabled: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.readOnlyDisabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    lightThemeWhiteBackground
+);
+
+export const interactions$DarkTheme$ReadOnlyAbsent$DisabledAbsent: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.none],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    darkThemeBlackBackground
+);
+
+export const interactions$DarkTheme$ReadOnlyAbsent$Disabled: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.disabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    darkThemeBlackBackground
+);
+
+export const interactions$DarkTheme$ReadOnly$DisabledAbsent: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.readOnly],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    darkThemeBlackBackground
+);
+
+export const interactions$DarkTheme$ReadOnly$Disabled: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.readOnlyDisabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    darkThemeBlackBackground
+);
+
+export const interactions$ColorTheme$ReadOnlyAbsent$DisabledAbsent: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.none],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    colorThemeDarkGreenBackground
+);
+
+export const interactions$ColorTheme$ReadOnlyAbsent$Disabled: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.disabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    colorThemeDarkGreenBackground
+);
+
+export const interactions$ColorTheme$ReadOnly$DisabledAbsent: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.readOnly],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    colorThemeDarkGreenBackground
+);
+
+export const interactions$ColorTheme$ReadOnly$Disabled: StoryFn = createFixedThemeStory(
+    createMatrixInteractions(component, [
+        [stepManipulationState.readOnlyDisabled],
+        stepLayoutStates,
+        selectedStates,
+        severityStates,
+        [stepContentStateShort, stepContentStateStepIndicator],
+    ]),
+    colorThemeDarkGreenBackground
+);
 
 export const hidden: StoryFn = createStory(
     hiddenWrapper(